pyproject.toml
[project]
name = "pybiscus"
version = "0.6.0"
description = "A flexible Federated Learning framework"
readme = "README.md"
requires-python = ">=3.10"
#authors = ["Baptiste MORISSE <baptiste.morisse@thalesgroup.com>"]
dependencies = [
    "einops>=0.8.0",
    "flask-cors>=5.0.1",
    "flask>=3.1.0",
    "flwr>=1.7.0",
    "lightning[pytorch-extra]>=2.4.0",
    "pandas>=2.2.3",
    "pydantic>=2.10.4",
    "pyyaml>=6.0.2",
    "torchvision>=0.20.1",
    "trogon>=0.6.0",
    "typer>=0.15.1",
    "wandb>=0.19.10",
    "pydeps>=3.0.1",
    "dominate>=2.9.1",
    "pillow>=11.0.0",
    "onnx>=1.18.0",
    "dotenv>=0.9.9",
]

# Optional dependencies (medical imaging)

[project.optional-dependencies]
medical_imaging = [
    "monai>=1.4.0",
    "nibabel>=5.0.0",
    "SimpleITK>=2.1.0",
    "lpips>=0.1.4",
]

[project.scripts]
pybiscus        = "pybiscus.main:main"
pybiscus_agent  = "pybiscus.session.agent.pybiscus_agent:main"
session_manager = "pybiscus.session.manager.start_manager:main"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["pybiscus"]

[dependency-groups]
test = [
    "pytest>=8.4.1",
    "pytest-xdist>=3.8.0",
]

